refactor into .py
