+++ src/relaxed/infer.py
@@ -11,7 +11,7 @@ import pyhf
 from chex import Array
 from jax import jit
 
-from ..mle import fit, fixed_poi_fit
+from relaxed.mle import fit, fixed_poi_fit
 
 
 def hypotest(
