--- src/relaxed/infer/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ("hypotest",)
-
-from relaxed.infer.hypothesis_test import hypotest
