--- src/relaxed/metrics/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ("asimov_sig", "gaussianity")
-
-from relaxed.metrics.likelihood_gaussianity import gaussianity
-from relaxed.metrics.significance import asimov_sig
