--- src/relaxed/mle/__init__.py
@@ -1,4 +1,0 @@
-__all__ = ("fit", "fixed_poi_fit")
-
-from relaxed.mle.constrained_fit import fixed_poi_fit
-from relaxed.mle.global_fit import fit
