--- src/relaxed/ops/__init__.py
@@ -1,10 +1,0 @@
-__all__ = (
-    "hist",
-    "fisher_info",
-    "cramer_rao_uncert",
-    "cut",
-)
-
-from relaxed.ops.cuts import cut
-from relaxed.ops.fisher_information import cramer_rao_uncert, fisher_info
-from relaxed.ops.histograms import hist
