+++ tests/test_metrics.py
@@ -2,11 +2,11 @@ import jax.numpy as jnp
 import numpy as np
 import pyhf
 import pytest
-from dummy_pyhf import example_model
 from jax import jacrev
 from jax.random import PRNGKey
 
 import relaxed
+from relaxed.dummy_pyhf import example_model
 
 
 @pytest.fixture
